Enable JS tests for resources library.
